feat(repo) create folder for joi schemas and middlewares, and refactor some code
