--- dbs/backup/dump.sql
@@ -1,1 +1,0 @@
-
