+++ docker-compose.yml
@@ -2,7 +2,7 @@ version: "3"
 services:
   mariadb:
     image: mariadb:latest
-    container_name: number8db
+    container_name: eventiodb
     ports:
       - 3306:3306
     environment:
@@ -13,4 +13,3 @@ services:
     volumes:
       - ./dbs/backup:/home/backup
       - ./dbs/init.sh:/docker-entrypoint-initdb.d/init.sh
-      - ./dbs/backup/dump.sql:/docker-entrypoint-initdb.d/dump.sql
